+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-18 11:48:49
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-08-18 13:47:32
+// Commit: feat(i18n): add async i18n stream provider
 
 class i18nComponentState {
-  final String commitRef = '3689';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-08-18T11:48:49+05:00');
+  final String commitRef = '3697';
+  final String lastMessage = 'feat(i18n): add async i18n stream provider';
+  final DateTime timestamp = DateTime.parse('2026-08-18T13:47:32+05:00');
 }