+++ CHANGELOG.md
@@ -3698,3 +3698,4 @@
 - [2026-08-18 13:34] feat(network): add async network stream provider
 - [2026-08-18 13:37] feat(test): add bi-directional event bus handler for test
 - [2026-08-18 13:45] feat(test): implement shimmer loading skeleton in test
+- [2026-08-18 13:47] feat(i18n): add async i18n stream provider